⚗️ Some very stupid ideas
