+++ Cargo.toml
@@ -24,8 +24,8 @@ tokio = { version = "1.37.0", features = ["io-util"], default-features = false, 
 async-mutex = { version = "1.4.0", optional = true }
 strum = { version = "0.26.2", features = ["derive"] }
 directories = "5.0.1"
-futures = { version = "0.3.31", optional = true }
-async-fs = { version = "2.1.2", optional = true }
+futures = { version = "0.3.31" }
+async-fs = { version = "2.1.2" }
 tokio-util = { version = "0.7.14", features = ["compat"], optional = true }
 
 [dev-dependencies]
@@ -36,9 +36,10 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [features]
-default = ["asynchronous"]
-asynchronous = ["dep:async-fs", "dep:async-mutex", "dep:futures"]
-tokio = ["asynchronous", "dep:tokio", "dep:tokio-util", "tokio/fs"]
+#default = ["asynchronous"]
+#asynchronous = ["dep:async-fs", "dep:async-mutex", "dep:futures"]
+blocking = []
+tokio = ["dep:tokio", "dep:tokio-util", "tokio/fs"]
 allow-unknown = []
 # Enables loading Journal files from before game version 4. Not fully supported and may lead to
 # parsing errors. In such cases, PRs with affected Journal files/events are appreciated :)
