+++ src/modules/io/models/log_dir/live_dir_iter.rs
@@ -1,25 +1,28 @@
-use crate::modules::io::{DirIter, LogError, LogFile, LogPath};
-use crate::modules::shared::blocking::sync_blocker::SyncBlocker;
+use crate::io::{DirIter, LogError, LogFile, LogPath};
+use crate::modules::shared::async_blocker::AsyncBlocker;
+use futures::{FutureExt, Stream};
 use notify::event::CreateKind;
 use notify::{Event, EventKind, RecommendedWatcher, RecursiveMode, Watcher};
 use std::collections::VecDeque;
 use std::path::Path;
+use std::pin::{pin, Pin};
 use std::sync::{Arc, Mutex};
+use std::task::{Context, Poll};
 
 #[derive(Debug)]
 pub struct LiveDirIter {
     dir_iter: DirIter,
-    blocker: SyncBlocker,
+    blocker: AsyncBlocker,
     last: Option<LogPath>,
     added: Arc<Mutex<VecDeque<Result<LogFile, LogError>>>>,
     _watcher: RecommendedWatcher,
 }
 
 impl LiveDirIter {
-    pub fn new<P: AsRef<Path>>(path: P) -> Result<LiveDirIter, LogError> {
-        let dir_iter = DirIter::new(path.as_ref())?;
+    pub async fn new<P: AsRef<Path>>(path: P) -> Result<LiveDirIter, LogError> {
+        let dir_iter = DirIter::new(path.as_ref()).await?;
 
-        let blocker = SyncBlocker::new();
+        let blocker = AsyncBlocker::default();
         let added = Arc::new(Mutex::new(VecDeque::new()));
 
         let local_blocker = blocker.clone();
@@ -27,21 +30,20 @@ impl LiveDirIter {
 
         // This is stopped when it is dropped
         let mut watcher = notify::recommended_watcher(move |event: notify::Result<Event>| {
-            dbg!(&event);
-
             let Ok(event) = event else {
                 return;
             };
 
-            if !matches!(event.kind, EventKind::Create(CreateKind::File) | EventKind::Create(CreateKind::Any)) {
-                dbg!("Returned");
+            if !matches!(event.kind, EventKind::Create(CreateKind::File)) {
                 return;
             }
 
-            let mut lock = local_added.lock().expect("lock should have been acquired");
+            let Ok(mut lock) = local_added.lock() else {
+                return;
+            };
 
             for path in event.paths {
-                let path = match dbg!(LogPath::try_from(path.as_path())) {
+                let path = match LogPath::try_from(path.as_path()) {
                     Ok(path) => path,
                     Err(LogError::IncorrectFileName) => continue,
                     Err(error) => {
@@ -55,8 +57,6 @@ impl LiveDirIter {
                 lock.push_back(Ok(log_file));
             }
 
-            dbg!("Unblocking...");
-
             local_blocker.unblock();
         })?;
 
@@ -65,8 +65,8 @@ impl LiveDirIter {
         Ok(LiveDirIter {
             dir_iter,
             blocker,
-            added,
             last: None,
+            added,
             _watcher: watcher,
         })
     }
@@ -78,17 +78,11 @@ impl LiveDirIter {
 
         log_file == last
     }
-}
 
-impl Iterator for LiveDirIter {
-    type Item = Result<LogFile, LogError>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(mut entry) = self.dir_iter.next() {
+    async fn inner_next(&mut self) -> Option<Result<LogFile, LogError>> {
+        if let Some(entry) = self.dir_iter.next() {
             self.last = Some(entry.log_path().clone());
-            entry.set_blocker(Arc::new(self.blocker.clone()));
-
-            // live_iter doesn't continue as the file doesn't know that it's not the last one.
+            // entry.set_blocker(Arc::new(self.blocker.clone()));
 
             return Some(Ok(entry));
         }
@@ -104,101 +98,83 @@ impl Iterator for LiveDirIter {
                 return added_value;
             }
 
-            self.blocker.block();
+            self.blocker.block().await;
         }
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use crate::modules::io::models::log_dir::live_dir_iter::LiveDirIter;
-    use crate::modules::shared::blocking::sync_blocker::SyncBlocker;
-    use crate::tests::test_dir;
-    use std::fs;
-    use std::thread::spawn;
-    use std::time::Instant;
-    use crate::modules::tests::simulate_log_dir;
+impl Stream for LiveDirIter {
+    type Item = Result<LogFile, LogError>;
 
-    #[test]
-    #[ignore]
-    fn live_watcher_blocks_correctly() {
-        let dir_path = simulate_log_dir("live_watcher_blocks_correctly");
-        let live_dir_iter = LiveDirIter::new(dir_path).unwrap();
-
-        let mut file_count = 0;
-
-        for file in live_dir_iter {
-            file_count += 1;
-
-            let file = file.unwrap();
-            let live_iter = file.live_iter().unwrap();
-
-            let mut i = 0;
-            let mut instant = Instant::now();
-            for entry in live_iter {
-                i += 1;
-                assert!(entry.is_ok());
-
-                dbg!(i);
-                dbg!(&entry);
-
-                // Simulation sleeps for 100 ms, so if ~100 ms have passed, we can be sure that the
-                // blocking has worked.
-                assert!(dbg!(instant.elapsed().as_millis()) > 90);
-
-                if i >= 75 {
-                    break;
-                }
-
-                instant = Instant::now();
-            }
-        }
-
-        unreachable!();
-
-        // let dir = test_dir();
-        // let first_file = dir.path().join("Journal.2023-02-21T084116.01.log");
-        // let second_file = dir.path().join("Journal.2023-02-21T084116.02.log");
-        //
-        // fs::write(
-        //     &first_file,
-        //     r#"{"timestamp":"2022-10-22T15:10:41Z","event":"Fileheader","part":1,"language":"English/UK","Odyssey":true,"gameversion":"4.0.0.1450","build":"r286858/r0 "}"#,
-        // )
-        //     .unwrap();
-        //
-        // let blocker = SyncBlocker::new();
-        //
-        // let local_blocker = blocker.clone();
-        // let local_path = dir.path();
-        //
-        // let handle1 = spawn(move || {
-        //     let mut live_dir = LiveDirIter::new(local_path).unwrap();
-        //     let mut file = live_dir.next().unwrap().unwrap().live_iter().unwrap();
-        //
-        //     assert!(file.next().is_some());
-        //
-        //     local_blocker.unblock();
-        //     assert!(file.next().is_none());
-        //
-        //     let mut next_file = live_dir.next().unwrap().unwrap().iter().unwrap();
-        //
-        //     assert!(next_file.next().is_some());
-        //     assert!(next_file.next().is_none());
-        //
-        //     assert!(true);
-        // });
-        //
-        // let handle2 = spawn(move || {
-        //     blocker.block();
-        //
-        //     fs::write(
-        //         &second_file,
-        //         r#"{"timestamp":"2022-10-22T15:10:41Z","event":"Fileheader","part":1,"language":"English/UK","Odyssey":true,"gameversion":"4.0.0.1450","build":"r286858/r0 "}"#,
-        //     )
-        //         .unwrap();
-        // });
-        //
-        // handle1.join().unwrap();
-        // handle2.join().unwrap();
+    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
+        pin!(self.inner_next()).poll_unpin(cx)
     }
 }
+
+// #[cfg(test)]
+// mod tests {
+//
+//     use crate::io::models::log_dir::async_live_dir_iter::AsyncLiveDirIter;
+//     use crate::modules::shared::async_blocker::AsyncBlocker;
+//     use crate::tests::test_dir;
+//     use futures::StreamExt;
+//     use smol::{fs, spawn};
+//
+//     // #[ignore]
+//     fn async_live_watcher_blocks_correctly() {
+//         smol::block_on(async {
+//             let dir = test_dir();
+//             let first_file = dir.path().join("Journal.2023-02-21T084116.01.log");
+//             let second_file = dir.path().join("Journal.2023-02-21T084116.02.log");
+//
+//             fs::write(
+//                 &first_file,
+//                 r#"{"timestamp":"2022-10-22T15:10:41Z","event":"Fileheader","part":1,"language":"English/UK","Odyssey":true,"gameversion":"4.0.0.1450","build":"r286858/r0 "}"#,
+//             )
+//                 .await
+//                 .unwrap();
+//
+//             let blocker = AsyncBlocker::new();
+//
+//             let local_blocker = blocker.clone();
+//             let local_path = dir.path();
+//
+//             let handle1 = spawn(async move {
+//                 let mut live_dir = AsyncLiveDirIter::new(local_path).await.unwrap();
+//                 let mut file = live_dir.next().await.unwrap().unwrap().live_iter().unwrap();
+//
+//                 assert!(file.next().is_some());
+//
+//                 local_blocker.unblock();
+//                 dbg!();
+//                 assert!(file.next().is_none());
+//                 dbg!();
+//
+//                 let mut next_file = live_dir.next().await.unwrap().unwrap().iter().unwrap();
+//
+//                 assert!(next_file.next().is_some());
+//                 assert!(next_file.next().is_none());
+//
+//                 assert!(true);
+//             });
+//
+//             let handle2 = spawn(async move {
+//                 dbg!("Handle 2 start");
+//                 blocker.block().await;
+//                 dbg!("Unblocked");
+//
+//                 fs::write(
+//                     &second_file,
+//                     r#"{"timestamp":"2022-10-22T15:10:41Z","event":"Fileheader","part":1,"language":"English/UK","Odyssey":true,"gameversion":"4.0.0.1450","build":"r286858/r0 "}"#,
+//                 )
+//                     .await
+//                     .unwrap();
+//
+//                 dbg!("Handle 2 done");
+//             });
+//
+//             handle1.await;
+//             handle2.await;
+//         });
+//     }
+// }
