--- src/modules/shared/asynchronous.rs
@@ -1,4 +1,0 @@
-pub mod async_blocker;
-
-#[cfg(feature = "tokio")]
-pub mod live_json_file_watcher;
