--- src/modules/shared/blocking.rs
@@ -1,2 +1,0 @@
-pub mod live_json_file_watcher;
-pub mod sync_blocker;
