+++ src/modules/shipyard/asynchronous.rs
@@ -3,8 +3,8 @@ use std::path::Path;
 use thiserror::Error;
 use tokio::fs;
 
-use crate::modules::shared::asynchronous::live_json_file_watcher::LiveJsonFileWatcher;
-pub use crate::modules::shared::asynchronous::live_json_file_watcher::LiveJsonFileWatcherError as ShipyardFileWatcherError;
+use crate::modules::shared::live_json_file_watcher::LiveJsonFileWatcher;
+pub use crate::modules::shared::live_json_file_watcher::LiveJsonFileWatcherError as ShipyardFileWatcherError;
 use crate::modules::shipyard::models::shipyard::Shipyard;
 
 pub type ShipyardFileWatcher = LiveJsonFileWatcher<Shipyard>;
